// src/types/generated/TransferRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TransferDirection } from "./TransferDirection";

export type TransferRequest = { serverId: string, direction: TransferDirection, localPath: string, remotePath: string, overwrite: boolean, };